+++ MetodosNumericos/src/herramientas/objetos/Modelo.cs
@@ -9,9 +9,9 @@ namespace MetodosNumericos.src.herramientas.objetos
     public abstract class Modelo
     {
         public int ValorN { get; set; }
-        public float ValorXinicial { get; set; }
+        public double ValorXinicial { get; set; }
         public Funcion Funcion { get; set; }
-        public abstract float[] resultados();
-        public abstract List<float> puntos();
+        public abstract double[] resultados();
+        public abstract List<double> puntos();
     }
 }
