+++ MetodosNumericos/src/herramientas/objetos/Unidad 4/ModeloTrapecio.cs
@@ -9,10 +9,10 @@ namespace MetodosNumericos.src.herramientas.objetos.Unidad4
 {
     public class ModeloTrapecio : Modelo
     {
-        public float intervaloA { get; set; }
-        public float intervaloB { get; set; }
+        public double intervaloA { get; set; }
+        public double intervaloB { get; set; }
 
-        public ModeloTrapecio(String funcion, float a, float b, int n)
+        public ModeloTrapecio(String funcion, double a, double b, int n)
         {
             base.Funcion = new Funcion(funcion);
             base.ValorN = n;
@@ -20,40 +20,40 @@ namespace MetodosNumericos.src.herramientas.objetos.Unidad4
             intervaloB = b;
         }
 
-        private float solucionSimple()
+        private double solucionSimple()
         {
-            float diferenciaIntervalos = intervaloB - intervaloA;
-            float sustitucionIntervaloA = base.Funcion.evaluar(this.intervaloA);
-            float sustitucionIntervaloB = base.Funcion.evaluar(this.intervaloB);
-            float resultado = diferenciaIntervalos * ((sustitucionIntervaloA + sustitucionIntervaloB) / 2);
+            double diferenciaIntervalos = intervaloB - intervaloA;
+            double sustitucionIntervaloA = base.Funcion.evaluar(this.intervaloA);
+            double sustitucionIntervaloB = base.Funcion.evaluar(this.intervaloB);
+            double resultado = diferenciaIntervalos * ((sustitucionIntervaloA + sustitucionIntervaloB) / 2);
             return resultado;
         }
 
-        private float solucionCompuesta()
+        private double solucionCompuesta()
         {
-            float diferenciaIntervalos = intervaloB - intervaloA;
-            float sustitucionIntervaloA = base.Funcion.evaluar(this.intervaloA);
-            float sumatoria = 0;
-            float intervalo = diferenciaIntervalos / (float)base.ValorN;
-            float acumulador = this.intervaloA + intervalo;
+            double diferenciaIntervalos = intervaloB - intervaloA;
+            double sustitucionIntervaloA = base.Funcion.evaluar(this.intervaloA);
+            double sumatoria = 0;
+            double intervalo = diferenciaIntervalos / (double)base.ValorN;
+            double acumulador = this.intervaloA + intervalo;
             while (acumulador != this.intervaloB)
             {
                 sumatoria += base.Funcion.evaluar(acumulador);
                 acumulador += intervalo;
             }
             
-            float sustitucionIntervaloB = base.Funcion.evaluar(this.intervaloB);
-            float resultado = diferenciaIntervalos * ((sustitucionIntervaloA + (2 * sumatoria) + sustitucionIntervaloB) / (2 * base.ValorN)) ;
+            double sustitucionIntervaloB = base.Funcion.evaluar(this.intervaloB);
+            double resultado = diferenciaIntervalos * ((sustitucionIntervaloA + (2 * sumatoria) + sustitucionIntervaloB) / (2 * base.ValorN)) ;
             return resultado;
         }
 
-        public override float[] resultados()
+        public override double[] resultados()
         {
-            float[] soluciones = new float[] { solucionSimple(), solucionCompuesta() }; 
+            double[] soluciones = new double[] { solucionSimple(), solucionCompuesta() }; 
             return soluciones;
         }
 
-        public override List<float> puntos()
+        public override List<double> puntos()
         {
             return base.Funcion.evaluar();
         }
