+++ MetodosNumericos/src/herramientas/objetos/Unidad 4/Resultados.cs
@@ -9,12 +9,12 @@ namespace MetodosNumericos.src.herramientas.objetos.Unidad_4
     public class Resultados
     {
         public String Funcion { get; set; }
-        public float DosPuntosInfinitasProgresivas { get; set; }
-        public float TresPuntosInfinitasProgresivas { get; set; }
-        public float DosPuntosInfinitasCentradas { get; set; }
-        public float TresPuntosInfinitasCentradas { get; set; }
-        public float DosPuntosInfinitasRegresivas { get; set; }
-        public float TresPuntosInfinitasRegresivas { get; set; }
+        public double DosPuntosInfinitasProgresivas { get; set; }
+        public double TresPuntosInfinitasProgresivas { get; set; }
+        public double DosPuntosInfinitasCentradas { get; set; }
+        public double TresPuntosInfinitasCentradas { get; set; }
+        public double DosPuntosInfinitasRegresivas { get; set; }
+        public double TresPuntosInfinitasRegresivas { get; set; }
         
         
     }
