+++ MetodosNumericos/src/herramientas/objetos/Unidad 4/ResultadosTrapecio.cs
@@ -9,7 +9,7 @@ namespace MetodosNumericos.src.herramientas.objetos.Unidad_4
     public class ResultadosTrapecio
     {
         public String Funcion { get; set; }
-        public float Simple { get; set; }
-        public float Compuesto { get; set; }
+        public double Simple { get; set; }
+        public double Compuesto { get; set; }
     }
 }
